Add platform properties
Here is a proof of concept how we could add specific
platform properties. This commit is not aimed to be merged,
but must be seen as how we could do this to clean #460,
and is to be evaluated.

If this looks convenient, then a clean proposal will be made,
without having the Microsoft specific properties in vexpress.
