+++ core/tee/tee_svc.c
@@ -169,11 +169,6 @@ static const uint32_t fw_impl_bin_version; /* 0 by default */
 /* Trusted firmware manufacturer name */
 static const char fw_manufacturer[] = TO_STR(CFG_TEE_FW_MANUFACTURER);
 
-#ifdef CFG_MICROSOFT_PROPERTIES
-/* Size of the TA endorsement seed */
-static const size_t ta_endorsement_seed_size = 32;
-#endif
-
 struct tee_props {
 	const void *data;
 	const size_t len;
@@ -197,22 +192,31 @@ const struct tee_props tee_props_lut[] = {
 	{fw_manufacturer, sizeof(fw_manufacturer)},
 	{0, 0}, /* client_id */
 	{0, 0}, /* ta_app_id */
-#ifdef CFG_MICROSOFT_PROPERTIES
-	{0, 0}, /* ta_endorsement_seed */
-#endif
+
+	/*
+	 * Properties specific to a given platform are not in the lut.
+	 * They are accessed through tee_svc_sys_get_property_platform()
+	 */
 };
+
+TEE_Result __weak tee_svc_sys_get_property_platform(uint32_t prop __unused,
+		tee_uaddr_t buf __unused, size_t blen __unused,
+		struct tee_ta_session *sess __unused)
+{
+	return TEE_ERROR_NOT_IMPLEMENTED;
+}
 
 TEE_Result tee_svc_sys_get_property(uint32_t prop, tee_uaddr_t buf, size_t blen)
 {
 	struct tee_ta_session *sess;
 	TEE_Result res;
 
-	if (prop > ARRAY_SIZE(tee_props_lut)-1)
-		return TEE_ERROR_NOT_IMPLEMENTED;
-
 	res = tee_ta_get_current_session(&sess);
 	if (res != TEE_SUCCESS)
 		return res;
+
+	if (prop > ARRAY_SIZE(tee_props_lut)-1)
+		return tee_svc_sys_get_property_platform(prop, buf, blen, sess);
 
 	switch (prop) {
 	case UTEE_PROP_TEE_DEV_ID:
@@ -271,55 +275,7 @@ TEE_Result tee_svc_sys_get_property(uint32_t prop, tee_uaddr_t buf, size_t blen)
 					    &sess->ctx->head->uuid,
 					    sizeof(TEE_UUID));
 
-#ifdef CFG_MICROSOFT_PROPERTIES
-	case UTEE_PROP_TA_ENDORSEMENT_SEED:
-		{
-			/*
-			 * The data to hash is 48 bytes made up of:
-			 * - 16 bytes: the UUID of the calling TA.
-			 * - 32 bytes: the hardware device ID
-			 * The resulting endorsement seed is 32 bytes.
-			 *
-			 * The output buffer is the "binary" struct defined in
-			 * the "prop_value" union and therefore comprises:
-			 * -  4 bytes: the size of the binary value data (32)
-			 * - 32 bytes: the binary value data (endorsement seed)
-			 *
-			 * Note that this code assumes an endoresement seed
-			 * size == device ID size for convenience.
-			 */
-			uint8_t data[sizeof(TEE_UUID)
-				+ ta_endorsement_seed_size];
 
-			uint8_t bin[sizeof(size_t)
-				+ ta_endorsement_seed_size];
-
-			size_t *bin_len = (size_t *) (void *) (&bin[0]);
-			uint8_t *bin_val = &bin[sizeof(size_t)];
-
-			if (blen < sizeof(bin))
-				return TEE_ERROR_SHORT_BUFFER;
-
-			memcpy(&data[0], &sess->ctx->head->uuid,
-				sizeof(TEE_UUID));
-
-			if (tee_otp_get_die_id(&data[sizeof(TEE_UUID)],
-									ta_endorsement_seed_size))
-				return TEE_ERROR_BAD_STATE;
-
-			res = tee_hash_createdigest(TEE_ALG_SHA256, data,
-					sizeof(data),
-					bin_val,
-					ta_endorsement_seed_size);
-			if (res != TEE_SUCCESS)
-				return TEE_ERROR_BAD_STATE;
-
-			*bin_len = ta_endorsement_seed_size;
-
-			return tee_svc_copy_to_user(sess, (void *)buf, bin,
-					sizeof(bin));
-		}
-#endif
 	default:
 		if (blen < tee_props_lut[prop].len)
 			return TEE_ERROR_SHORT_BUFFER;
