+++ lib/libutee/sub.mk
@@ -13,3 +13,8 @@ srcs-y += tee_api_operations.c
 srcs-y += tee_api_se.c
 
 subdirs-y += arch/$(ARCH)
+
+ifeq ($(CFG_PLATFORM_SPECIFIC_PROPERTIES),y)
+global-incdirs-y += arch/$(ARCH)/plat-$(PLATFORM)
+srcs-y += arch/$(ARCH)/plat-$(PLATFORM)/platform_properties.c
+endif
